% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch_eval_cor.R
\name{batch_eval_cor}
\alias{batch_eval_cor}
\title{Batch Evaluation and Correction}
\usage{
batch_eval_cor(
  exprFile,
  batchFile,
  batch,
  NameString = "",
  discrete.batch = TRUE
)
}
\arguments{
\item{exprFile}{Path to .txt (tab-delimited) or a .csv file that contains the expression data. In the file, rows
should be genes and columns should be sample IDs.}

\item{batchFile}{Path to the .txt (tab-delimited) or a .csv file that contains the batch information.
The first column should be sample names/IDs}

\item{batch}{The title of the batch for which you want to evaluate and do the
correction. Default = "Batch"}

\item{NameString}{String that will be added in to all output filenames. Default="".}

\item{discrete.batch}{Logical value, indicating whether the samples are already
grouped in discrete batches. If the value is FALSE, contiguous batch information is
clustered into discrete batches using mClust. Useful for clustering batch variables that are
contiguous. Default = TRUE}
}
\description{
Evaluates if the batch effects with a gene expression dataset using linear
regression and if the batch is associated with the data, the data is batch-corrected
using the ComBat Algorithm.

Please set your working directory before you call the function as all output files will be saved to this folder.
Ensure that you add the correct input file paths.
}
\examples{
\dontrun{
batch_eval_cor(exprFile = "~/exprData.txt",
               batchFile = "~/batchData.txt",
               Batch = "Batch",
               discrete.batch = TRUE)
}


}
