% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxplot_data.R
\name{boxplot_data}
\alias{boxplot_data}
\title{Boxplots showing Gene Expression before and after correction}
\usage{
boxplot_data(expr1, expr2, NameString, batch = "Batch")
}
\arguments{
\item{NameString}{String that will be added in all output filenames. Default=NA.}

\item{batch}{The title of the batch for which you want to evaluate and do the
correction. Default = "Batch"

@import grDevices

@export}

\item{exp1}{A dataframe containing gene expression values before batch correction.
Rows are genes, columns are samples.}

\item{exp2}{A dataframe containing gene expression values after adjustment
for batch effects. #' Rows are genes, columns are samples.}
}
\description{
Plots boxplots showing the change in gene expression after
adjustment for batch effects using ComBat
}
