% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlationPlot.R
\name{correlationPlot}
\alias{correlationPlot}
\title{Correlation Scatter Plot}
\usage{
correlationPlot(data1, data2, fileName = "", batch)
}
\arguments{
\item{data1}{A matrix containing the gene expression data before batch correction}

\item{data2}{A matrix containing the gene expression data after batch correction}

\item{fileName}{String to be added to the name of the file to which correlation
scatter plot would be saved}

\item{batch}{The tile of the batch}
}
\description{
The function calculates the correlation coefficient using
the expression data before and after batch correction and also plots a scatter
plot to show the correlation.
}
