% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmeans_PCA.R
\name{kmeans_PCA}
\alias{kmeans_PCA}
\title{PCA plot with k-means information}
\usage{
kmeans_PCA(exprData, batch.info, batch = "Batch", NameString = "", when)
}
\arguments{
\item{exprData}{gene expression dataset (rows should be samples, column should be genes)}

\item{batch.info}{contains the samples names and the batches they belong to}

\item{batch}{title of the batch being used for correction}

\item{NameString}{string that will be appear in all output filenames. Default="" (empty string)}

\item{when}{String indicating when the clustering is taking place (before batch correction or after batch correction?)}
}
\value{
Returns the optimal number of clusters (k) that has the maximum average silhouette width
and was used for clustering
}
\description{
Clusters the data using k-means after finding the optimum number
of clusters using the silhouette method and displays the results on a PCA plot
saved to a file. The files containing the avg. silhouette width (for k = 2 to
k = 7) and the cluster information re saved to the kmeans folder within the working directory
}
