% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMF_PCA.R
\name{NMF_PCA}
\alias{NMF_PCA}
\title{PCA plot with NMF cluster information}
\usage{
NMF_PCA(expr, batch.info, nrun = 30, batch = "Batch", NameString = "", when)
}
\arguments{
\item{expr}{gene expression dataset (rows should be genes, column should be samples)}

\item{batch.info}{contains the samples names and the batches they belong to}

\item{nrun}{The number of runs for NMF, the default number is 30. The consensus matrices from the NMF results are used to compute the cophenetic coefficients}

\item{batch}{title of the batch being used for correction}

\item{NameString}{string that will be appear in all output filenames. Default="" (empty string)}

\item{when}{String indicating when the clustering is taking place (before batch correction or after batch correction?)}
}
\value{
Returns the optimal number of clusters (k) that has the maximum
average silhouette width (ignoring k=2) and was used for clustering and plotting.
}
\description{
Clusters the data using NMF (Non-neagtive matrix factorization)
after finding the optimal number of clusters in a dataset using the value of
cophenetic coefficients.The results of the clustering are used along with
PCA to see whether all the samples of a batch lie in the same cluster.
The cophenetic coefficient plot, the PCA biplot, and the files containing the
cophentic coefficeients (for number of clusters: 2 to 7) and the clustering
information (for optimal k) is saved to the NMF folder created in the working
directory.
}
