% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca_batch.R
\name{pca_batch}
\alias{pca_batch}
\title{PCA Plot}
\usage{
pca_batch(data, batch.info, batch, plotFile)
}
\arguments{
\item{data}{A matrix containing gene expression data. Row names should be
samples and column names should be genes.}

\item{batch.info}{A data frame containing the samples names and details of the
batch they belong to.}

\item{batch}{Title of the batch the data is being corrected for.}

\item{plotFile}{Name of the file in which the PCA is to be plotted.

@import grDevices
@import stats
@import graphics}
}
\description{
Plots the first two principal components for all the samples of
a gene expression dataset along with the batch information
}
