% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca_prop_var.R
\name{pca_prop_var}
\alias{pca_prop_var}
\title{PCA Proportion of Variation}
\usage{
pca_prop_var(batch.title, plotFile, data1, data2)
}
\arguments{
\item{batch.title}{The tile of the batch}

\item{plotFile}{The PCA proportion of variation will be saved to this file}

\item{data1}{A matrix containing the gene Expression data before batch correction}

\item{data2}{A matrix containing the gene Expression data after batch correction}
}
\description{
Calculates and plots the proportion of variation for twenty-five Principal Components of
expression data before and after batch correction
}
