% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beacon.R
\name{beacon}
\alias{beacon}
\title{Batch Effects Evaluation and Correction}
\usage{
beacon(
  expr1,
  batch.info,
  batch,
  NameString = "",
  discrete.batch = TRUE,
  clus.method = "both",
  nrun.NMF = 30
)
}
\arguments{
\item{expr1}{gene expression data; rows should be genes and columns should be samples.}

\item{batch.info}{contains the batch information. The first column should be sample names.}

\item{batch}{The title of the batch for which you want to evaluate and do the
correction. Default = "Batch"}

\item{NameString}{String that will be added in all output filenames. Default=NA.}

\item{discrete.batch}{Logical value indicating whether the samples are already
grouped in discrete batches. If the value is FALSE, contiguous batch information is
clustered into discrete batches. Useful for clustering batch variables that are
contiguous, for example, used reads and useful reads in mClust. Default = TRUE}

\item{clus.method}{Method to be used for clustering. "km" denotes k-means, "NMF"
denotes NMF with 30 runs. Default means clustering is done using both the methods.}

\item{nrun.NMF}{number of runs for NMF. Default = 30.}
}
\description{
Evaluates if the batch effects with a gene expression dataset using linear
regression and if the batch is associated with the data, the data is batch-corrected
using the ComBat Algorithm.

Please set your working directory before you call the function as all output files will be saved to this folder.
}
\examples{
\dontrun{
beacon(expr1 = dataset,
       batch.info = batch_info,
       batch = "Batch",
       discrete.batch = TRUE)
}


}
