% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ComBat_data.R
\name{ComBat_data}
\alias{ComBat_data}
\title{Batch Correction using ComBat}
\usage{
ComBat_data(expr, batch.info, batch = "Batch", NameString = "")
}
\arguments{
\item{expr}{gene expression data to be corrected for batch effects. Rows
should be genes and columns should be samples.}

\item{batch.info}{A data frame containing batch information corresponding to
the gene expression dataset,the first column should be sample names and the
second column should contain the batch the sample belongs to.}

\item{batch}{The title of the batch for which you want to evaluate and do the
correction. Default = "Batch"}

\item{NameString}{String that will be added to the names of all output files. Default=NA.}
}
\value{
Returns a matrix containing the batch corrected values for gene expression.
}
\description{
Corrects the gene expression data for a particular batch using
ComBat algorithm
}
