% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lin_reg.R
\name{lin_reg}
\alias{lin_reg}
\title{Linear Regression of Principal Components}
\usage{
lin_reg(data, batch.info, batch = "Batch")
}
\arguments{
\item{data}{A matrix containing gene expression data. Row names should be
samples and column names should be genes.}

\item{batch.info}{A data frame containing the samples names and details of the
batch they belong to.}

\item{batch}{Title of the batch the data is being corrected for.}
}
\value{
Returns the p-value after linear regression analysis if the batch is
associated with data.
}
\description{
Calculates Linear Regression of Principal Components 1 and 2 of the data
to see if the batch is correlated to the data.
}
