% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca_batch.R
\name{pca_batch}
\alias{pca_batch}
\title{PCA Plot with batch information}
\usage{
pca_batch(exprData, batch.info, batch, NameString = "", when = "")
}
\arguments{
\item{exprData}{A matrix containing gene expression data. Row names should be
samples and column names should be genes.}

\item{batch.info}{A data frame containing the samples names and details of the
batch they belong to.}

\item{batch}{Title of the batch the data is being corrected for.}

\item{NameString}{string that will be appear in all output filenames. Default="" (empty string)}

\item{when}{String indicating for which dataset is the PCA plot being created

@import grDevices
@import stats
@import graphics}
}
\description{
Plots the first two principal components for all the samples of
a gene expression dataset along with the batch information
}
